test(eval): cover edge cases, move impl out of index for coverage
Lines coverage 93.82% → 100% (threshold 95%).

- Move snapshot/index.ts and diff/index.ts impl into sibling files; barrel re-exports remain. Shared vitest config excludes index.ts, which was hiding the new modules from the metric.

- Add tests for parseCassette version/entries errors, fingerprintRequest context branch, lastUserContent no-user path, loose miss, sequential exhaustion, replay abort no-op.
